# apps/alerts-service/openapi.yaml
openapi: 3.0.3
info:
  title: Alerts Service API
  version: 0.1.0
servers:
  - url: https://alerts.example.com
paths:
...
  /api/notifications/send:
    post:
      summary: Queue a notification (and send immediately if SMTP configured)
      security:
        - internalKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendRequest'
      responses:
        '200':
          description: Queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
        '400': { description: Invalid payload }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    internalKey:
      type: apiKey
      in: header
      name: x-internal-key
  schemas:
    SendRequest:
      type: object
      required: [channel, to, subject]
      properties:
        channel: { type: string, enum: ['email'] }
        to: { type: string, format: email }
        subject: { type: string }
        text: { type: string }
        html: { type: string }
        tags:
          type: array
          items: { type: string }
        scheduledAt: { type: string, format: date-time }
    SendResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            id: { type: string, format: uuid }
...